PhysicalPolyhedron commit

+++ src/geometry/Tetrahedron.java
@@ -42,10 +42,11 @@ public class Tetrahedron {
     }
 
     public double getVolume(){
+
         final Vector3D a = new Vector3D(D, B);
         final Vector3D b = new Vector3D(D, A);
         final Vector3D c = new Vector3D(D, C);
-
-        return (double) ((1/6) * a.scalarProduct(b.vectorProduct(c)));
+        
+        return (1d/6d) * a.scalarProduct(b.vectorProduct(c));
     }
 }
